Order 기능 추가
-주문생성
-주문 리스트 조회
-특정 주문 조회
-주문수정
-주문상태 업데이트+ 취소기능

+++ src/main/java/com/example/commerce_back_office/controller/ProductController.java
@@ -45,7 +45,7 @@ public class ProductController {
      * - 존재하지 않는 ID 요청 시 예외 발생
      */
     @GetMapping("/{id}")
-    public ResponseEntity<CommonResponse<ProductDetailResponse>> getProduct(@PathVariable Integer id) {
+    public ResponseEntity<CommonResponse<ProductDetailResponse>> getProduct(@PathVariable Long id) {
 
         ProductDetailResponse product = productService.getProduct(id);
         return ResponseEntity.status(OK).body(CommonResponse.of(GET_PRODUCT, product));
@@ -76,7 +76,7 @@ public class ProductController {
     @PatchMapping("/{id}")
 
     public ResponseEntity<CommonResponse<ProductDetailResponse>> updateProduct(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @Valid @RequestBody ProductUpdateRequest request
     ) {
         ProductDetailResponse updatedProduct = productService.updateProduct(id, request);
